+++ lib/vagrant/action/box/destroy.rb
@@ -10,8 +10,12 @@ module Vagrant
         end
 
         def call(env)
+          # Delete the existing box
           env[:ui].info I18n.t("vagrant.actions.box.destroy.destroying", :name => env[:box_name])
           FileUtils.rm_rf(env[:box_directory])
+
+          # Reload the box collection
+          env[:box_collection].reload!
 
           @app.call(env)
         end